# complete-api-docs.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/users/{id}":  {
                                          "get":  {
                                                      "operationId":  "UsersController_getUserById",
                                                      "summary":  "Get user by ID",
                                                      "description":  "Retrieves detailed user information by user ID",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "id",
                                                                             "required":  true,
                                                                             "in":  "path",
                                                                             "description":  "User ID",
                                                                             "schema":  {
                                                                                            "example":  "user-123-uuid-456",
                                                                                            "type":  "string"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "User retrieved successfully"
                                                                                },
                                                                        "401":  {
                                                                                    "description":  "Unauthorized"
                                                                                },
                                                                        "404":  {
                                                                                    "description":  "User not found"
                                                                                }
                                                                    },
                                                      "tags":  [
                                                                   "Users"
                                                               ],
                                                      "security":  [
                                                                       {
                                                                           "bearer":  [

                                                                                      ]
                                                                       }
                                                                   ]
                                                  }
                                      },
                  "/api/users/telegram/{telegramId}":  {
                                                           "get":  {
                                                                       "operationId":  "UsersController_getUserByTelegramId",
                                                                       "summary":  "Get user by Telegram ID",
                                                                       "description":  "Retrieves user information using their Telegram ID",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "telegramId",
                                                                                              "required":  true,
                                                                                              "in":  "path",
                                                                                              "description":  "Telegram user ID",
                                                                                              "schema":  {
                                                                                                             "example":  "123456789",
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "User retrieved successfully"
                                                                                                 },
                                                                                         "401":  {
                                                                                                     "description":  "Unauthorized"
                                                                                                 },
                                                                                         "404":  {
                                                                                                     "description":  "User not found"
                                                                                                 }
                                                                                     },
                                                                       "tags":  [
                                                                                    "Users"
                                                                                ],
                                                                       "security":  [
                                                                                        {
                                                                                            "bearer":  [

                                                                                                       ]
                                                                                        }
                                                                                    ]
                                                                   }
                                                       },
                  "/api/users/{id}/stats":  {
                                                "get":  {
                                                            "operationId":  "UsersController_getUserStats",
                                                            "summary":  "Get user statistics",
                                                            "description":  "Retrieves comprehensive learning statistics for a user including progress, streaks, and performance metrics",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "required":  true,
                                                                                   "in":  "path",
                                                                                   "description":  "User ID",
                                                                                   "schema":  {
                                                                                                  "example":  "user-123-uuid-456",
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "User statistics retrieved successfully"
                                                                                      },
                                                                              "401":  {
                                                                                          "description":  "Unauthorized"
                                                                                      },
                                                                              "404":  {
                                                                                          "description":  "User not found"
                                                                                      }
                                                                          },
                                                            "tags":  [
                                                                         "Users"
                                                                     ],
                                                            "security":  [
                                                                             {
                                                                                 "bearer":  [

                                                                                            ]
                                                                             }
                                                                         ]
                                                        }
                                            },
                  "/api/users/{id}/settings":  {
                                                   "get":  {
                                                               "operationId":  "UsersController_getUserSettings",
                                                               "summary":  "Get user settings",
                                                               "description":  "Retrieves user preferences and configuration settings",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "required":  true,
                                                                                      "in":  "path",
                                                                                      "description":  "User ID",
                                                                                      "schema":  {
                                                                                                     "example":  "user-123-uuid-456",
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "User settings retrieved successfully"
                                                                                         },
                                                                                 "401":  {
                                                                                             "description":  "Unauthorized"
                                                                                         },
                                                                                 "404":  {
                                                                                             "description":  "User not found"
                                                                                         }
                                                                             },
                                                               "tags":  [
                                                                            "Users"
                                                                        ],
                                                               "security":  [
                                                                                {
                                                                                    "bearer":  [

                                                                                               ]
                                                                                }
                                                                            ]
                                                           },
                                                   "post":  {
                                                                "operationId":  "UsersController_updateUserSettings",
                                                                "summary":  "Update user settings",
                                                                "description":  "Updates user preferences and configuration settings",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "id",
                                                                                       "required":  true,
                                                                                       "in":  "path",
                                                                                       "description":  "User ID",
                                                                                       "schema":  {
                                                                                                      "example":  "user-123-uuid-456",
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "requestBody":  {
                                                                                    "required":  true,
                                                                                    "description":  "Settings to update",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "object",
                                                                                                                                            "properties":  {
                                                                                                                                                               "notifications":  {
                                                                                                                                                                                     "type":  "boolean",
                                                                                                                                                                                     "example":  true
                                                                                                                                                                                 },
                                                                                                                                                               "language":  {
                                                                                                                                                                                "type":  "string",
                                                                                                                                                                                "example":  "en"
                                                                                                                                                                            },
                                                                                                                                                               "timezone":  {
                                                                                                                                                                                "type":  "string",
                                                                                                                                                                                "example":  "UTC"
                                                                                                                                                                            },
                                                                                                                                                               "studyReminders":  {
                                                                                                                                                                                      "type":  "boolean",
                                                                                                                                                                                      "example":  true
                                                                                                                                                                                  }
                                                                                                                                                           }
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Settings updated successfully"
                                                                                          },
                                                                                  "400":  {
                                                                                              "description":  "Invalid settings data"
                                                                                          },
                                                                                  "401":  {
                                                                                              "description":  "Unauthorized"
                                                                                          },
                                                                                  "404":  {
                                                                                              "description":  "User not found"
                                                                                          }
                                                                              },
                                                                "tags":  [
                                                                             "Users"
                                                                         ],
                                                                "security":  [
                                                                                 {
                                                                                     "bearer":  [

                                                                                                ]
                                                                                 }
                                                                             ]
                                                            }
                                               },
...
                  "/api/admin/users/{id}":  {
                                                "get":  {
                                                            "operationId":  "AdminController_getUser",
                                                            "summary":  "Get user details (admin only)",
                                                            "description":  "Retrieves detailed information about a specific user including admin-only data",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "required":  true,
                                                                                   "in":  "path",
                                                                                   "description":  "User ID",
                                                                                   "schema":  {
                                                                                                  "example":  "user-123-uuid-456",
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "User details retrieved successfully"
                                                                                      },
                                                                              "401":  {
                                                                                          "description":  "Unauthorized"
                                                                                      },
                                                                              "403":  {
                                                                                          "description":  "Forbidden - Admin access required"
                                                                                      },
                                                                              "404":  {
                                                                                          "description":  "User not found"
                                                                                      }
                                                                          },
                                                            "tags":  [
                                                                         "Admin"
                                                                     ],
                                                            "security":  [
                                                                             {
                                                                                 "bearer":  [

                                                                                            ]
                                                                             }
                                                                         ]
                                                        }
                                            },
...
              },
    "info":  {
                 "title":  "Matrus API",
                 "description":  "AI-powered Leitner System API",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
...
                   }
}